LPS-200222 Adapting publication, now it is simpler
